mercadolibre/requirements.txt
appdirs==1.4.4
attrs==20.3.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
black==20.8b1
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
decorator==4.4.2
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
itemadapter==0.2.0
itemloaders==1.0.4
jedi==0.18.0
jmespath==0.10.0
lxml==4.6.2
mypy-extensions==0.4.3
parsel==1.6.0
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.14
Protego==0.1.16
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.7.4
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
queuelib==1.5.0
regex==2020.11.13
requests==2.25.1
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
soupsieve==2.1
toml==0.10.2
traitlets==5.0.5
Twisted==20.3.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
w3lib==1.22.0
wcwidth==0.2.5
zope.interface==5.2.0
